+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Improved /etc/shadow file parsing for the nxUser resource.
+
+## [1.5.0] - 2025-01-28
+
+### Fixed
+
 - Fixed module not working in PowerShell v7.4.x due to nxFileSystemInfo not implementing the abstract property Exists from the base type FileSystemInfo.
 - Fixed nxService reporting wrong status on Ubuntu Server 22.04 LTS due to incorrect parsing.
 - Fixed nxGroup not removing group members when specifying that the group should be empty.